tests: fix ga orders unit tests.

+++ pallets/genetic-analysis-orders/src/mock.rs
@@ -84,7 +84,7 @@ impl pallet_timestamp::Config for Test {
 
 impl pallet_randomness_collective_flip::Config for Test {}
 
-type Balance = u64;
+type Balance = u128;
 
 parameter_types! {
 	pub static ExistentialDeposit: Balance = 0;
@@ -162,7 +162,7 @@ impl user_profile::Config for Test {
 }
 
 pub struct ExternalityBuilder {
-	existential_deposit: u64,
+	existential_deposit: u128,
 }
 
 impl Default for ExternalityBuilder {
@@ -172,7 +172,7 @@ impl Default for ExternalityBuilder {
 }
 
 impl ExternalityBuilder {
-	pub fn existential_deposit(mut self, existential_deposit: u64) -> Self {
+	pub fn existential_deposit(mut self, existential_deposit: u128) -> Self {
 		self.existential_deposit = existential_deposit;
 		self
 	}
